requirements.txt
pandas
flask
faiss-cpu
argparse
markdown2
openai
python-dotenv
numpy